removed two superfluous tests and moved spec.js files to correct location
